poc/requirements.txt
streamlit
langchain
langchain-openai
langchain-anthropic
langchain-chroma
langchain-community
python-dotenv
